Added progress based on bytes copied, made it performing a move only on the same drive, fixed bugs.

+++ operations/Addition.cs
@@ -8,8 +8,9 @@ namespace Kopi
 {
     class Addition
     {
-        public Addition(string a_sourcePath, string a_destinationPath)
+        public Addition(Copyer.Context a_context, string a_sourcePath, string a_destinationPath)
         {
+            m_context = a_context;
             FileInfo fileInfo = new FileInfo(a_sourcePath);
             SourcePath = a_sourcePath;
             DestinationPath = a_destinationPath;
@@ -20,11 +21,16 @@ namespace Kopi
 
         public void Execute()
         {
-            // We don't overwrite because if the destination file existed then it would be a Modification rather than an Addition.
-            Directory.CreateDirectory(Path.GetDirectoryName(DestinationPath));
-            File.Copy(SourcePath, DestinationPath, false);
+            m_context.LogDelegate("Copying new file " + SourcePath + " to " + DestinationPath);
+            if (!m_context.DryRun)
+            {
+                // We don't overwrite because if the destination file existed then it would be a Modification rather than an Addition.
+                Directory.CreateDirectory(Path.GetDirectoryName(DestinationPath));
+                File.Copy(SourcePath, DestinationPath, false);
+            }
         }
 
+        private Copyer.Context m_context;
         public string Name { get; set; }
         public string SourcePath { get; set; }
         public string DestinationPath { get; set; }
